src/main/java/nu/lansingcarworkshop/services/converters/LocalDateConverter.java
package nu.lansingcarworkshop.services.converters;

import javax.persistence.AttributeConverter;
import javax.persistence.Converter;
import java.sql.Date;
import java.time.LocalDate;

@Converter(autoApply = true)
public class LocalDateConverter implements AttributeConverter<LocalDate, Date> {

    @Override
    public Date convertToDatabaseColumn(LocalDate localDate) {
        return (localDate == null ? Date.valueOf(LocalDate.now()) : Date.valueOf(localDate));
    }

    @Override
    public LocalDate convertToEntityAttribute(Date date) {
        return (date == null ? LocalDate.now() : date.toLocalDate());
    }

}